+++ packages/web/components/CreateOrgOnboarding/index.ts
@@ -1,0 +1,1 @@
+export * from "./CreateOrgOnboarding";
