+++ packages/web/components/Dashboard/DashboardHeader/index.ts
@@ -1,0 +1,1 @@
+export * from "./DashboardHeader";
