+++ packages/web/components/Dashboard/DashboardNavbar/index.ts
@@ -1,0 +1,1 @@
+export * from "./DashboardNavbar";
