+++ packages/web/components/LatestCommits/index.ts
@@ -1,1 +1,2 @@
 export * from "./LatestCommits";
+export * from "./Commit";
