+++ packages/web/components/PageShell/PageShell.tsx
@@ -1,4 +1,4 @@
-import { Container, AppShell } from "@mantine/core";
+import { AppShell, type AppShellProps } from "@mantine/core";
 import { useQuery } from "@tanstack/react-query";
 import axios from "axios";
 import { useRouter } from "next/router";
@@ -8,10 +8,10 @@ import { PageLoader } from "../PageLoader";
 import { getLoggedOutPageHeader } from "./utils";
 
 type PageShellProps = {
-  children: React.ReactNode;
+  appShellProps: AppShellProps;
 };
 
-export const PageShell: React.FC<PageShellProps> = ({ children }) => {
+export const PageShell: React.FC<PageShellProps> = ({ appShellProps }) => {
   const router = useRouter();
   const loggedOutPageHeader = getLoggedOutPageHeader(router.pathname);
 
@@ -38,5 +38,5 @@ export const PageShell: React.FC<PageShellProps> = ({ children }) => {
     );
   }
 
-  return <Container>{children}</Container>;
+  return <AppShell {...appShellProps} />;
 };
